# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ARGWS Scout API
  version: 0.7.0
  description: REST API para espaços de trabalho, coleta web e publicação via Connect|API, com canais e instâncias WhatsApp isolados logicamente.
servers:
  - url: /v1
security:
  - ScoutSession: []
  - ScoutToken: []
paths:
...
  /jobs/{jobId}:
    get:
      parameters:
        [
          {
            in: path,
            name: jobId,
            required: true,
            schema: { type: string, format: uuid },
          },
        ]
      summary: Retorna estado, resultado, source e tentativas.
      responses:
        "200": { description: Job encontrado. }
...
components:
  securitySchemes:
    ScoutSession: { type: apiKey, in: cookie, name: scout_access }
    ScoutToken: { type: http, scheme: bearer, bearerFormat: Scout API token }
    PreAuth: { type: http, scheme: bearer, bearerFormat: MFA pre-auth token }
...